+++ src/org/labkey/test/BaseWebDriverTest.java
@@ -187,9 +187,6 @@ public abstract class BaseWebDriverTest extends LabKeySiteWrapper implements Cle
     protected static boolean _checkedLeaksAndErrors = false;
     private static final String ACTION_SUMMARY_TABLE_NAME = "actions";
 
-    protected static final String PERMISSION_ERROR = "User does not have permission to perform this operation.";
-    protected static final String NOT_FOUND_ERROR = "notFound";
-
     static final Set<String> urlsSeen = new HashSet<>();
 
     static
@@ -1541,12 +1538,6 @@ public abstract class BaseWebDriverTest extends LabKeySiteWrapper implements Cle
         {
             log("properties were already set, no changed needed");
         }
-    }
-
-    public void assertAtUserUserLacksPermissionPage()
-    {
-        assertTextPresent(PERMISSION_ERROR);
-        assertTitleEquals("403: Error Page -- User does not have permission to perform this operation.");
     }
 
     public void assertNavTrail(String... links)
